chore: move tests to own folder
avoid compiling unwanted files
