+++ jest.config.ts
@@ -1,4 +1,6 @@
-module.exports = {
+import type {Config} from 'jest'
+
+const config: Config = {
   preset: 'ts-jest',
   clearMocks: true,
   noStackTrace: true,
@@ -10,4 +12,7 @@ module.exports = {
   },
   testEnvironment: 'jest-environment-jsdom',
   verbose: true, // Set to false to see console log during tests.
+  testPathIgnorePatterns: ['/node_modules/', '/lib/', '/documentation/'],
 }
+
+export default config
