+++ package.json
@@ -5,8 +5,8 @@
   "homepage": "https://github.com/AoDev/bard-router",
   "main": "lib/index.js",
   "scripts": {
-    "build:js": "del-cli lib && babel src --out-dir lib --extensions '.ts,.tsx' --ignore '**/*.spec.ts'",
-    "build:types": "tsc --emitDeclarationOnly",
+    "build:js": "del-cli lib && babel src --out-dir lib --extensions '.ts,.tsx'",
+    "build:types": "tsc --emitDeclarationOnly --listEmittedFiles",
     "build": "npm run build:js && npm run build:types",
     "lint": "eslint src --quiet",
     "prepare": "NODE_ENV=production && npm run build",
@@ -67,6 +67,7 @@
     "react-dom": "18.2.0",
     "standard-version": "9.5.0",
     "ts-jest": "29.1.1",
+    "ts-node": "10.9.1",
     "typescript": "5.2.2"
   }
 }
