+++ tsconfig.json
@@ -1,5 +1,4 @@
 {
-  "exclude": ["node_modules", "**/*.spec.ts"],
   "include": ["./src/**/*"],
   "compilerOptions": {
     "target": "esnext",
